refactor: Split true types and data models into seperate files
